Polishing before I switch the dongle to back - Adjusted symbol layer - moved out the totem layout to another file so I can do other keyboards - cleaned out old macros and defines - go back to balanced on the urob timerless
