+++ config/totem.h
@@ -1,0 +1,60 @@
+/*
+ * Position definition for timerless home row mods
+ *
+ *   ╭─────────────────────────┬─────────────────────────╮
+ *   │     LT4 LT3 LT2 LT1 LT0 │ RT0 RT1 RT2 RT3 RT4     │
+ *   │     LM4 LM3 LM2 LM1 LM0 │ RM0 RM1 RM2 RM3 RM4     │
+ *   │ LB5 LB4 LB3 LB2 LB1 LB0 │ RB0 RB1 RB2 RB3 RB4 RB5 │
+ *   ╰───────────╮ LH2 LH1 LH0 │ RH0 RH1 RH2 ╭───────────╯
+ *               ╰─────────────┴─────────────╯
+ */
+
+#define LT4 0
+#define LT3 1
+#define LT2 2
+#define LT1 3
+#define LT0 4
+
+#define RT0 5
+#define RT1 6
+#define RT2 7
+#define RT3 8
+#define RT4 9
+
+#define LM4 10
+#define LM3 11
+#define LM2 12
+#define LM1 13
+#define LM0 14
+
+#define RM0 15
+#define RM1 16
+#define RM2 17
+#define RM3 18
+#define RM4 19
+
+#define LB5 20
+#define LB4 21
+#define LB3 22
+#define LB2 23
+#define LB1 24
+#define LB0 25
+
+#define RB0 26
+#define RB1 27
+#define RB2 28
+#define RB3 29
+#define RB4 30
+#define RB5 31
+
+#define LH2 32
+#define LH1 33
+#define LH0 34
+
+#define RH0 35
+#define RH1 36
+#define RH2 37
+
+#define KEYS_L LT4 LT3 LT2 LT1 LT0 LM4 LM3 LM2 LM1 LM0 LB5 LB4 LB3 LB2 LB1 LB0
+#define KEYS_R RT0 RT1 RT2 RT3 RT4 RM0 RM1 RM2 RM3 RM4 RB0 RB1 RB2 RB3 RB4 RB5
+#define THUMBS LH2 LH1 LH0 RH0 RH1 RH2
